HOTFIX: check offset limits in streamtask when recovering KTable store
